requirements.txt
elasticsearch==2.2.0
django-allauth==0.24.1
djangorestframework-xml==1.3.0
#apt-xapian-index==0.45
argparse==1.2.1
#chardet==2.0.1
colorama==0.2.5
configobj==4.7.2
Django==1.8.9
djangorestframework==3.3.2
django-filter==0.12.0
funcy==1.2
html5lib==0.999
httplib2==0.9
jsonschema==2.4.0
lxml==3.4.1
Markdown==2.6.5
mock==1.0.1
MySQL-python==1.2.5
oauth2==1.5.211
#PAM==0.4.2
#pyOpenSSL==0.13
pyserial==2.6
pyzmq==14.5.0
requests==2.2.1
shortuuid==0.4.2
simplejson==3.6.4
South==1.0.2
ssh-import-id==3.21
stevedore==0.15
tornado==4.1
urllib3==1.7.1
virtualenv-clone==0.2.5
virtualenv==1.11.6
virtualenvwrapper==4.3.1
wsgiref==0.1.2
zope.interface==4.0.5
djangorestframework-jwt==1.7.2
django-rest-auth==0.7.0
django-extensions==1.6.1
django-super-inlines==0.1.4
celery==3.1.24
django-celery==3.1.17
jsonfield==1.0.3
Pillow==4.0.0
djangorestframework-yaml==1.0.3
sqlparse==0.2.3
supervisor==3.2.0
redis==2.10.3
